Cleanly separate the selection listener
