+++ src/presence/highlights/Listener.tsx
@@ -8,27 +8,13 @@ import "rangy/lib/rangy-selectionsaverestore";
 
 const rangy = (rangyCore as any).default;
 
-export default function SelectionListener() {
-  // Store the nodes that are enabled for highlights
-  const [containers, setContainers] = useState<Record<string, Element>>({});
-  const [serializedSelection, setSerializedSelection] = useState<string | null>(
-    null
-  );
-
-  // Build containers
-  useEffect(() => {
-    // get all nodes that have a data-highlights attribute
-    const enabled = document.querySelectorAll("[data-highlights]");
-    const newContainers: Record<string, Element> = {};
-    for (const element of enabled) {
-      const checksum: string = rangy.getElementChecksum(element);
-      newContainers[checksum] = element;
-    }
-    //console.log(newContainers);
-    setContainers(newContainers);
-    (document as any).rangy = rangy;
-  }, []);
-
+export default function Listener({
+  containers,
+  setSelection,
+}: {
+  containers: Record<string, Element>;
+  setSelection: (selection: string | null) => void;
+}) {
   const getContainer = (selection: RangySelection) => {
     for (const range of selection.getAllRanges()) {
       for (const element of Object.values(containers)) {
@@ -56,12 +42,12 @@ export default function SelectionListener() {
       // @ts-ignore
       const selection = rangy.getSelection();
       if (selection.isCollapsed) {
-        setSerializedSelection(null);
+        setSelection(null);
         return;
       }
       const container = getContainer(selection);
       if (!(container && isEntirelyContained(selection, container))) {
-        setSerializedSelection(null);
+        setSelection(null);
         return;
       }
       const serialized = rangy.serializeSelection(
@@ -69,7 +55,7 @@ export default function SelectionListener() {
         false, // omit the container element
         container
       );
-      setSerializedSelection(serialized);
+      setSelection(serialized);
     };
     document.addEventListener("selectionchange", handleSelectionChange);
 
@@ -77,10 +63,6 @@ export default function SelectionListener() {
       document.removeEventListener("selectionchange", handleSelectionChange);
     };
   }, [containers]);
-
-  useEffect(() => {
-    console.log("serializedSelection", serializedSelection);
-  }, [serializedSelection]);
 
   return null;
 }
